Update DataNode struct and add pubsub flags
This commit prepares for introduction of a new way to configure publication
channels via flags in the data nodes instead of arrays with node IDs.

Unfortunately, the change increases size of each node by 4 bytes on 32-bit
systems, even though only 2 bytes would be needed. Instead of wasting the
padded bytes, the access variable was also increased to 16-bit.