Replace bin_sub with bin_import function
Rationale: The import function should be used to deal with data that's not
directly coming from the comms interface, but e.g. from non-volatile
memory storage.

As a consequence, the new import function does not contain any ThingSet
function byte or a path/endpoint. It is the direct equivalent to the
bin_export function to get serialized data back into the data objects.

Subscription to statements should be implemented in the process function
instead of a dedicated function.

bin_sub has been deprecated and will be removed in future releases.