Allow more fine-grained selection of features
The following features can now be switched on or off via ts_config.h or
Kconfig (if using Zephyr):
- 64-bit types (uint64_t, int64_t, double) support
- CBOR decimal fraction support
- CBOR byte string support

Two unit-tests are now run: One with minimal feature set and one with
all above features enabled.

Also fixing issue of previous code which checked features using #ifdef
instead of #if, which was always true as the features are either defined
as 0 or 1. Thus, this commit should lead to significant code footprint
improvements for standard configuration.