Fix deprecated C++ pub/sub functions
The txt/bin_pub and bin_sub functions are now provided as inlined
functions in the C++ interface only.

They were not used in the new C-based interface anyway and deprecation
of the C interface caused warnings even if the functions were not used
by the application anymore, as they were always called by the C++ shim
layer.

Unit-tests for the deprecated functions had to be removed because they
can only use the C interface.