COM prepare - add ThingSet OSAL for communication buffers
Prepare for ThingSet Communication Context:
- Add operating system abstraction for communication buffers.
- Add Native implementation of communication buffers
- Add Zephyr implementation of communication buffers

Communication buffers are a core concept of how ThingSet passes
communication data around.

Every device holds a pool of buffers that can be utilized on the device.
The buffers and the buffers pool are modeled after (or uses) the Zephyr
network buffers. By using a single buffer pool the buffers can easily be
transfered to and from any ThingSet context.

Signed-off-by: Bobby Noelte <b0661n0e17e@gmail.com>