Add nested JSON support for statements
The ThingSet specification v0.5 introduces a different data layout compared to
previous versions where the data is grouped by entities of the device (like
battery, actuator) instead of the data type (e.g. measurement, configuration).
The data type is described by a single character prefix in the item name.
The new grouping allows to have same item names in different groups, so for
unambigous description of the data the nested structure has to be maintained in
statements.

This option is introduced to maintain compatibility with legacy firmware and
will be enabled by default in the future.