Use flag to store if item was updated
Previously, configuration values had to be stored all in a single category
with a callback, which could be used to store data in the NVM after an update.

If data items in all groups can be stored values (as planned for ThingSet v0.5)
the callback method is not feasible anymore. Instead, the new flag can be used
to identify if items were updated.

Currently it is not checked if the new value was actually different from the
previously stored value. Any patch request received via ThingSet will set
the updated flag of all included items to true.

As we are now changing the data object, some const specifiers had to be
removed.

The number of allowed subsets in the bit set had to be reduced to 7 to keep the
overall memory footprint.