Add update notification via callback from patch functions
This is a new attempt for the reverted method using a single update flag
per data object.

Prior to v0.5 spec, configuration values had to be stored all in a single
group with a callback assigned to it. This callback could be used to store
data in the NVM after an update.

If data items in all groups can be stored values (as planned for ThingSet v0.5)
the group callback method is not feasible anymore. Instead, an update
notification callback per patch request as introduced in this commit is
necessary.

Currently it is not checked if a new value was actually different from the
previously stored value. Any valid patch request received via ThingSet will
call the specified update callback function.