Add iterable macros for records (BREAKING CHANGE)
The macros for iterable records and record items were still missing.

Unfortunately, for Zephyr iterables to work, the first item has to be
a unique ID, so we need to change the order of the existing TS_RECORD_ITEM_*
macros.

The order was changed such that the compiler won't compile code with the
old order, so that the breaking change is obvious and. However, a breaking
change cannot be avoided.

Advantage: The order is now more aligned with existing macros, where the
unique ID is first and the parent ID is behind the pointer or the detail
field.